+++ lib/commands/fortnite/locker/locker.dart
@@ -1,6 +1,7 @@
 import "package:nyxx_commands/nyxx_commands.dart";
 import "text.dart";
 import "image.dart";
+import "full.dart";
 import "exclusives.dart";
 import "crew.dart";
 import "special.dart";
@@ -11,6 +12,7 @@ final ChatGroup lockerCommand = ChatGroup(
   children: [
     lockerTextCommand,
     lockerImageCommand,
+    lockerFullImageCommand,
     lockerExclusivesImageCommand,
     lockerCrewImageCommand,
     lockerSpecialImageCommand,