+++ lib/commands/fortnite/stw/inventory/inventory.dart
@@ -3,6 +3,7 @@ import "package:nyxx_commands/nyxx_commands.dart";
 import "backpack.dart";
 import "storage.dart";
 import "ventures.dart";
+import "backpack_and_storage.dart";
 
 final ChatGroup inventoryCommand = ChatGroup(
   "inventory",
@@ -11,6 +12,7 @@ final ChatGroup inventoryCommand = ChatGroup(
     backpackInventoryCommand,
     storageInventoryCommand,
     venturesInventoryCommand,
+    backpackAndStorageInventoryCommand,
   ],
   checks: [],
 );